Added Web Script to support object/user loading from Web Framework
A single web script that grabs a content object and hands back the metadata as JSON.  This is for a single object and is used in various places.  One such place is during the Login Process.  The AlfrescoUserFactory calls over to retrieve the user's properties, which it can then make available to web scripts running within the Web Tier.

The author was very careful not to rely upon json.js or to generate JSON via Javascript.  The much-esteemed and greatly praised doclib example was closely followed and the results work quite well.


git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@9070 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261