SLNG-1252
Earlier the the repo webscripts for wiki, blog, forum/discussions and comments stripped the userinput (content field) from un-safe html tags such as scripts. Only white listed tags was allowed (specified in StringUtils.safeTags). Now <strong> and <em> has been added to the list but the method to strip away the tags, StringUtils.stripUnsafeHTML() is no longer used from these scripts.

Instead the share webtier is doing the cleaning of these tags when they are rendered using a html.ftl file. (this is the case for the wiki)
If they are requested from the client directly through an ajax proxy call the result is stripped using the Alfresco.util.stripUnsafeHTML in alfresco.js. (this is the case for blog, forum/discussions and comments).

base.css has been updated to display "rich content", in other words user input that has been formatted, correctly.

A new mimetype for installing air applications has also been added to the list of mime types.

git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@11003 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261