ALF-10117: JBPM workflows should be hidden.
A new property "system.workflow.jbpm.definitions.visible" has been added, set to false by default. This means that when the JBPM engine is enabled (the default for upgrades) in-flight workflows can continue but the JBPM definitions will not be returned and therefore hide them from the UI preventing new instances from being created.

If customers want to show the JBPM workflows again all is required is to set the "system.workflow.jbpm.definitions.visible" property to true.

The Workflow MBean has also been updated to expose the new property and the Workflow admin console now displays the value.

git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@30356 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261