RM-2027 Refactor mock authentication code.
Create a new class to handle creation of the mock authentication util.
It would be nice to always use it as a factory to create a mock util, but
unfortunately this causes many of the existing unit tests to fail. Something
clever is happening in org.mockito.MockitoAnnotations.initMocks(Object) that
I currently don't understand.

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@100292 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261