RM-2947 Remove unnecessary chemistry dependency.
This dependency is provided by the dataprep librarywhich also provides
the implementation. Both should be version 0.10.0 or 0.11.0, but we
were taking 0.10.0 of one and 0.11.0 of the other. This should fix the
NoSuchMethodErrors we were seeing.

(cherry picked from commit 9d510a46c9db392a8abe0158f530afe6f7a1984a)