RM-2812 Add audit service and dependencies.
Having thought about this more, I don't think there's any problem adding
a deprecated class to the public API, as long as it's clearly marked as
such. It's deprecated at the moment because we think people might be using
it, but we want them to migrate to something else. Declaring this as part of
the public API formalises this process.

The RecordsManagementAuditService has the following dependencies:

RecordsManagementAuditQueryParameters <- RecordsManagementAuditService
RecordsManagementAuditEntry <- RecordsManagementAuditService
AuditEvent <- RecordsManagementAuditService