RM-3344 Ensure we are not downgrading admin or upgrading guest.
This is a first pass at preventing changes to the clearance of admin or guest.
Previously we weren't allowing any changes to their clearance, but when
working with security markings, admin in particular needs to occasionally gain
a new mark when a custom group is created.  Currently that clearing is not
happening, but this change does not prevent it, if needed.

Also remove an unused method from the CaveatMarkingServiceImpl, as the class
is already very full.