RM-4159 Wrap the RestWrapper rather than extending it.
Previously we were hitting an issue where there were three beans all of the
type RestWrapper. This meant that Spring could not determine which bean
should be used when autowiring.  By wrapping the RestWrapper we ensure there
is only one bean of that type, and so Spring knows which bean to use (in the
TAS code).  After some debate I decided against renaming the RMRestWrapper
as RestWrapperWrapper.