RM-5481 Handle empty string as a reason or exemption id.
If the empty string is passed as a classification reason or exemption id then the API framework
converts it to null for us. We should throw a 404 (not found) error, because there are no reasons
or exemptions with the empty string as their id (or null for that matter).  Add two new REST API
tests for these cases.

Also stopped using the helper method addParam in TopicsWithInstructionsTest, as it is only saving us
about ten characters, causes an extra import and includes an explicit mention of the topicsApi
inside it (i.e. it doesn't work at all with the guidesAPI).