MNT-18546: exclude the undeletable aspect from node copies
Added a policy implementation to the UndeletableAspect class so that
the aspect is not copied to other nodes.

This solves the problem where a document has the aspect added to it by
Cloud sync (for example) and is then checked-out, but may never be
checked in again (or editing cancelled) as the working copy included
the sys:undeletable aspect.