REPO-2575: added property loader pattern for non-enterprise projects
Now that we have "decomposed" various parts of the mega-monolith repository
into smaller projects such as alfresco-remote-api, we need a mechanism for
adding new system properties within the project that requires them.

For example, to add a new boolean/toggle for the WWW-Authenticate behaviour
in the remote API project, we don't want that setting to go in
alfresco-repository - we want it to live in alfresco-remote-api. So this
extra pattern (classpath*:alfresco/project.properties) is being added to
load these project specific properties.