Fix/repo 2025 cat tests (#31)
REPO-2025 Marked existing tests with categories LuceneTests and DBTests

Also includes a few PerformanceTests. So that the DB tests could be run on their own for each DB, a new
AllDBTestsTestSuite was created and it was tried out with MySQL (REPO-2002), MariaDB (REPO-2005) and PostgreSQL (REPO-2954).

A few existing TestSuites have been removed because we now have AllDBTestsTestSuite or the same set was covered by one of the TestSuites run in the latest Bamboo build jobs.