MNT-18894 UI Threads blocked at getCurrentRepositoryDescriptor
DescriptorServiceImpl modified so that the synchronization blocks used to ensure that license checks on startup run in the correct order don't result in hangs in the UI later when UI requests result in getCurrentRepositoryDescriptor() calls. The code have been modified so that there still are synchronization blocks when the license changes, but the access to the
currentRepositoryDescriptor is no longer included in the same block, but protected by a separate ReadWriteLock.