REPO-2253/ALF-21854: Action parameter lookup for "de_DE" falls back to "root" locale instead of "de"
Form labels must fall back to being shown in the language of the
specified locale when there is no country specific text provided.

For example if de-DE (Germany) is requested, but there is no bundle
for Germany specifically, just the German language then we should still
get German back, and not en-US or whatever the server is defaulting to.