REPO-3744 Make query limit configurable (#167)
The hardcoded 100 max items limit was changed to a configurable parameter authority.findAuthorityLimit which now has a default of 10000.
This limit is applied to a query that is going in SearchService, therefore can go to Solr instead of database if Full Text Search queries are configured to use Solr instead of the database.
As the query returns all items and sorting and paging is done in remote-api in GroupsImpl#getAuthoritiesInfo, the execution can be slow on large datasets. Local tests showed about 6s to get results for a group of 10000 users.