REPO-4114  POST /nodes/{nodeId}/children api does not use Transform Service for renditions (#114)
In addition to using the Transform service for renditions when uploading content via the v1 REST api, this PR also includes changes to allow a client to request multiple renditions when uploading content or to request multiple renditions in a single request. This is now possible as the newer RenditionService2 is async.