RM-6141: RM_support_for_the_Insight_Engine (#301)
* RM-6141: RM_support_for_the_Insight_Engine

Pass filter queries in SQL requests to the Insight Engine.

* SEARCH-1351 Add unit tests for SolrSQLHttpClient.

This involved changing the return type of postSolrQuery. The cast
taking place in this file did not make sense to me, since the two
types are not in the same type hierarchy. However it is obviously
working at the moment, so I have simply moved the location of the
cast. I couldn't find any references to the protected method anyway,
so it did not seem risky to update.

* SEARCH-1351 Add new unit test file to suite.

* SEARCH-1351 Remove unused test variables and fix logging.