REPO-4639 Content conversion failed using Tika (#587)
The Tika T-Engine was not being called by the Transform Service or Local transforms, because the sub-transform name was required.

Moved common classes (with repo) to alfresco-transform-model so that it can be used by the T-Engines to work out which sub transformer to use. InT also performed a refactor on these classes (alfresco-transform-model 1.0.2.7). Also includes changes to legacy transformers so that they don't pass the sub transformer name any more.