REPO-4663: Error accessing Admin Console as non literal admin (#382)
* Run as system while authentication

Authentication to the console as system to get the perms
for non literal admin.

* Test when run as system while authentication

Authentication to the console as system to get the permissions
for a non literal admin user. Basic Auth is test.

* Test: Clear System runner before authentication

Run authentication as user not System to make the fix act

* Test: Modifications after PR comments

Usage of this removed, unnecessary
Rid of delete user because it slows down test execution.
Wrapping the authentication process to give a clearer message.