WIP: POC REPO-4791 If possible, use Local transforms where Legacy transforms are called. (#388)
Linked to https://github.com/Alfresco/alfresco-repository/pull/611

The objective of this PR is to replace all the ContentService calls related to transforms with calls to the SynchronousTransformClient that switches between Local and Legacy transforms. This will allow us to remove the Legacy transforms at some point and also take advantage of the T-Engines which can be scaled.