MNT-22409 Legacy transformer fails to generate multi-page TIFF preview
The reason for this bug is quite complex, but is solvable.

    The original pdf rendition (defined in ACS 5.2 using XML) incorrectly used a SWFTransformationOptions object rather than an ImageTransformationOptions object.

    In an previous version of ACS a workaround had been added to ImageMagickContentTransformerWorker if the transform options was not a ImageTransformationOptions object. It used a pageLimit property if set on the transform option to limit the number of pages read. In ACS 6.2.2 this is now 1, resulting in only one page being included in the preview. The command to ImageMagick includes [0] after the source file name (as can be seen in a previous comment).

    In ACS 6.2.2 the pageLimit is being set to 1, because transform options needed to be converted backwards and forwards between the new and old formats by the TransformOptionsConverter. In ACS 6.1.1 the SWFTransformationOptions object had an unlimited pageLimit value.

    The pageLimit was being set to 1, because the default system.thumbnail.definition.default.pageLimit property value was 1. This is one of a group of unfortunately named properties that include the word thumbnail. A thumbnail in Alfresco means all renditions including the preview and not just the small image renditions.

So the cleanest way to fix this is to do the following:

    Add startPage and endPage transform options to the renditions that produce small images. All of which use ImageMagick. These are the transform option names are known to T-Engines that performs the ImageMagick transforms and also the Legacy ImageMagickContentTransformerWorker.

    Change the system.thumbnail.definition.default.pageLimit to be -1 (unlimited) so that it is not added to transform options simply because they are being converted.
    There is a risk this may impact custom rendition that expect the pageLimit to be set to 1 by default, but this requires a small change in the custom rendition if it is a problem.

    Fix up the automated test around the TransformOptionsConverter.

Note the Legacy transformers have been removed in ACS 7.0.0, so this issue is not seen. However the changes to the rendition definitions and system.thumbnail.definition.default.pageLimit will be added to future Major, Minor or Service Pack releases.  Changes will be made to 6.2.N and 7.0.N and on request to hot fix branches.

While looking into this, it turns out that the T-Engines also have a bug. The code was refactored at some point and a new ImageMagickOptionsBuilder class was created. The startPage and endPage transform options are accepted but just ignored. This explains why when the system is using Local transforms or the transform service multiple pages are seen in the preview. A separate bug will be raised for this. The impact might be that ImageMagick is doing more work than it needs to. It is also possible that the thumbnail option always included in the renditions for these small images is being used to avoid this.